super quick refactor for the demo tomorrow
